Endpointler yetkilere göre düzenlendi.
