+++ src/main/java/tr/edu/ogu/ceng/controller/CompanySupervisorController.java
@@ -21,7 +21,9 @@ import lombok.AllArgsConstructor;
 import tr.edu.ogu.ceng.dto.CompanySupervisorDto;
 import tr.edu.ogu.ceng.dto.requests.CompanySupervisorRequestDto;
 import tr.edu.ogu.ceng.dto.responses.CompanySupervisorResponseDto;
+import tr.edu.ogu.ceng.enums.InternshipStatus;
 import tr.edu.ogu.ceng.service.CompanySupervisorService;
+import tr.edu.ogu.ceng.service.InternshipService;
 import tr.edu.ogu.ceng.util.PageableUtil;
 
 @RestController
@@ -30,6 +32,7 @@ import tr.edu.ogu.ceng.util.PageableUtil;
 public class CompanySupervisorController {
 	@Autowired
 	private final CompanySupervisorService service;
+	InternshipService internshipService;
 
 	@GetMapping
 	public Page<CompanySupervisorResponseDto> getAll(@RequestParam(defaultValue = "0") Integer pageNo,
@@ -68,5 +71,20 @@ public class CompanySupervisorController {
 	@GetMapping("/getCompanySupervisorByUserId/{userId}")
 	public CompanySupervisorDto getCompanySupervisorByUserId(@PathVariable Long userId) {
 		return service.getCompanySupervisorByUserId(userId);
+	}
+
+	@PutMapping("/company-approved/{id}")
+	public InternshipStatus approveInternshipByCompany(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.COMPANY_APPROVED);
+	}
+
+	@PutMapping("/company-evaluation-stage/{id}")
+	public InternshipStatus moveToInternshipEvaluationStageByCompany(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.COMPANY_EVALUATION_STAGE);
+	}
+	
+	@PutMapping("/company-rejected/{id}")
+	public InternshipStatus rejectInternshipByCompany(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.COMPANY_REJECTED);
 	}
 }
