+++ src/main/java/tr/edu/ogu/ceng/controller/FacultySupervisorController.java
@@ -20,12 +20,14 @@ import tr.edu.ogu.ceng.dto.requests.FacultySupervisorRequestDto;
 import tr.edu.ogu.ceng.dto.requests.StudentRequestDto;
 import tr.edu.ogu.ceng.dto.responses.FacultySupervisorResponseDto;
 import tr.edu.ogu.ceng.dto.responses.StudentResponseDto;
+import tr.edu.ogu.ceng.enums.InternshipStatus;
 import tr.edu.ogu.ceng.enums.UserType;
 import tr.edu.ogu.ceng.model.FacultySupervisor;
 import tr.edu.ogu.ceng.model.Student;
 import tr.edu.ogu.ceng.model.User;
 import tr.edu.ogu.ceng.service.EmailService;
 import tr.edu.ogu.ceng.service.FacultySupervisorService;
+import tr.edu.ogu.ceng.service.InternshipService;
 import tr.edu.ogu.ceng.service.StudentService;
 import tr.edu.ogu.ceng.util.PageableUtil;
 import tr.edu.ogu.ceng.util.PasswordGeneratorUtil;
@@ -42,6 +44,7 @@ public class FacultySupervisorController {
 
 	@Autowired
 	StudentService studentService;
+	InternshipService internshipService;
 
 	@GetMapping("/supervisors")
 	public Page<FacultySupervisorResponseDto> getAllFacultySupervisors(@RequestParam(defaultValue = "0") Integer pageNo,
@@ -98,5 +101,40 @@ public class FacultySupervisorController {
 	@GetMapping("/byUserId/{userId}")
 	public FacultySupervisorResponseDto getFacultySupervisorByUserId(@PathVariable Long userId) {
 		return facultySupervisorService.getFacultySupervisorByUserId(userId);
+	}
+
+	@PutMapping("/faculty-approved/{id}")
+	public InternshipStatus approveInternshipByFaculty(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.FACULTY_APPROVED);
+	}
+	
+	@PutMapping("/ongoing/{id}")
+	public InternshipStatus markInternshipAsOngoing(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.ONGOING);
+	}
+
+	@PutMapping("/faculty-evaluation-stage/{id}")
+	public InternshipStatus moveToInternshipEvaluationStageByFaculty(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.FACULTY_EVALUATION_STAGE);
+	}
+	
+	@PutMapping("/faculty-rejected/{id}")
+	public InternshipStatus rejectInternshipByFaculty(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.FACULTY_REJECTED);
+	}
+
+	@PutMapping("/faculty-invalid/{id}")
+	public InternshipStatus markInternshipAsInvalidByFaculty(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.FACULTY_INVALID);
+	}
+
+	@PutMapping("/success/{id}")
+	public InternshipStatus markInternshipAsSuccess(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.SUCCESS);
+	}
+
+	@PutMapping("/canceled/{id}")
+	public InternshipStatus cancelInternship(@PathVariable(name = "id") long id) {
+	    return internshipService.chanceInternshipStatus(id, InternshipStatus.CANCELED);
 	}
 }
